FAC-711: Move findActiveMembers query to ExpertGroupAccessJpaRepository
